pyproject.toml
[project]
name = "market_inteligence_crew"
version = "0.1.0"
description = "market_inteligence_crew using crewAI"
authors = [{ name = "Sameema Tariq", email = "sameema.tariq.ds@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.14.2",
    "langchain-cerebras>=0.1.0",
    "langchain>=0.2.0",
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1"
]

[project.scripts]
market_inteligence_crew = "market_inteligence_crew.main:run"
run_crew = "market_inteligence_crew.main:run"
train = "market_inteligence_crew.main:train"
replay = "market_inteligence_crew.main:replay"
test = "market_inteligence_crew.main:test"
run_with_trigger = "market_inteligence_crew.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
